Water quality monitoring data from Yarra River (Melbourne, Australia)

A dataset containing water quality measurements from monitoring sites across Yarra River (Melbourne, Australia), including various chemical and physical parameters measured between 1991 and 2025

Usage

yarra_water

Format

A data frame with 1,501 rows and 13 variables:

site_id

Factor. Unique identifier for the monitoring site (e.g., 229143)

name

Factor. Name of the monitoring site location (e.g., "YARRA @ CHANDLER HWY")

datetime

POSIXct. Date and time when the sample was collected

data_type

Factor. Type of data collected - 2 levels: Quality and Quantity)

parameter

Factor. The water quality parameter measured - 22 levels (e.g., "Total Arsenic")

value

Numeric. The measured value of the parameter

unit_of_measurement

Character. The unit in which the parameter is measured (e.g., "mg/L")

year

Numeric. Year when the sample was collected

month

Numeric. Month when the sample was collected (1-12)

weekday

Ordered factor. Day of the week when the sample was collected

period

Factor. Decade period - 2 levels: 1990s and 2020s)

season

Factor. Season when the sample was collected - 4 levels: Autumn, Winter, Spring, Summer

parameter_type

Factor. Category of the parameter - 3 levels: Toxicant, Nutrient, Physical/Chemical Stressor

Source

Victorian Department of Energy, Environment, and Climate Action https://data.water.vic.gov.au/WMIS

Examples

# Load the dataset
data(yarra_water)

# View structure
str(yarra_water)
#> tibble [1,501 × 13] (S3: tbl_df/tbl/data.frame)
#>  $ site_id            : Factor w/ 5 levels "229108","229142",..: 3 3 3 3 3 3 3 3 3 3 ...
#>  $ name               : Factor w/ 5 levels "YARRA @ CHANDLER HWY",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ datetime           : POSIXct[1:1501], format: "2020-04-02 08:55:00" "2020-06-05 12:20:00" ...
#>  $ data_type          : Factor w/ 2 levels "Quality","Quantity": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ parameter          : Factor w/ 22 levels "Total Arsenic",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ value              : num [1:1501] 0.001 0.001 0.003 0.001 0.001 0.001 0.001 0.001 0.001 0.001 ...
#>  $ unit_of_measurement: chr [1:1501] "mg/L" "mg/L" "mg/L" "mg/L" ...
#>  $ year               : num [1:1501] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 ...
#>  $ month              : num [1:1501] 4 6 8 9 9 9 9 9 9 9 ...
#>  $ weekday            : Ord.factor w/ 7 levels "Sun"<"Mon"<"Tue"<..: 5 6 2 4 5 6 7 1 2 3 ...
#>  $ period             : Factor w/ 2 levels "1990s","2020s": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ season             : Factor w/ 4 levels "Spring","Summer",..: 3 4 4 1 1 1 1 1 1 1 ...
#>  $ parameter_type     : Factor w/ 3 levels "Physical/Chemical Stressor",..: 3 3 3 3 3 3 3 3 3 3 ...

# Summary statistics
summary(yarra_water)
#>    site_id                       name         datetime                  
#>  229108:   0   YARRA @ CHANDLER HWY:1501   Min.   :1991-01-08 13:50:00  
#>  229142:   0   YARRA @ TEMPLESTOWE :   0   1st Qu.:1994-07-27 11:00:00  
#>  229143:1501   YARRA @ WARRANDYTE  :   0   Median :2022-01-10 13:00:00  
#>  229147:   0   YARRA @ YERING GORGE:   0   Mean   :2014-11-30 11:36:11  
#>  229200:   0   YARRA DS MCMAHONS   :   0   3rd Qu.:2024-02-21 10:50:01  
#>                                            Max.   :2025-03-26 10:50:00  
#>                                                                         
#>     data_type                parameter       value          unit_of_measurement
#>  Quality :1501   Water Temperature:225   Min.   :   0.001   Length:1501        
#>  Quantity:   0   pH               :223   1st Qu.:   0.130   Class :character   
#>                  Salinity as EC@25:222   Median :   8.700   Mode  :character   
#>                  Turbidity        :218   Mean   :  41.224                      
#>                  Total Arsenic    : 53   3rd Qu.:  26.000                      
#>                  Total Chromium   : 53   Max.   :1300.000                      
#>                  (Other)          :507                                         
#>       year          month        weekday     period        season   
#>  Min.   :1991   Min.   : 1.000   Sun: 11   1990s: 412   Spring:373  
#>  1st Qu.:1994   1st Qu.: 3.000   Mon:129   2020s:1089   Summer:481  
#>  Median :2022   Median : 6.000   Tue:195                Autumn:327  
#>  Mean   :2014   Mean   : 6.215   Wed:715                Winter:320  
#>  3rd Qu.:2024   3rd Qu.: 9.000   Thu:251                            
#>  Max.   :2025   Max.   :12.000   Fri:189                            
#>                                  Sat: 11                            
#>                     parameter_type
#>  Physical/Chemical Stressor:964   
#>  Nutrient                  :240   
#>  Toxicant                  :297   
#>                                   
#>                                   
#>                                   
#>